+++ CHANGELOG.rst
@@ -4,9 +4,11 @@ Changelog
 Version 0.1.8
 -------------
 - Adds missing cerebellum regions in the direction vectors calculation and refactors the implementation [`NSETM_1682_`]
+- Cell counts' linear program: adds no constraints when estimates are missing [`NSETM-1689`_]
+- Uses a voxel-based computation of placement hints as an alternative to meshes [`NSETM-1660`_] [`NSETM-1670`_]
 
-Version 0.1.7
--------------
+Version 0.1.7 (2021-11-8)
+-------------------------
 - Zero negative density values when they are known to be round-off errors [`NSETM_1665_`]
 
 Version 0.1.6 (2021-10-29)
@@ -81,7 +83,9 @@ Version 0.1.0 (2021-04-27)
 - CLI to compute a flat map based on the streamlines of a laminar brain region
 - CLI to turn the excel measurement compilation gaba_papers.xlsx of D. Rodarie into a CSV file.
 
-
+.. _`NSETM-1689`: https://bbpteam.epfl.ch/project/issues/browse/NSETM-1689
+.. _`NSETM-1670`: https://bbpteam.epfl.ch/project/issues/browse/NSETM-1670
+.. _`NSETM-1660`: https://bbpteam.epfl.ch/project/issues/browse/NSETM-1660
 .. _`NSETM-1682`: https://bbpteam.epfl.ch/project/issues/browse/NSETM-1682
 .. _`NSETM-1665`: https://bbpteam.epfl.ch/project/issues/browse/NSETM-1665
 .. _`CA-20`: https://bbpteam.epfl.ch/project/issues/browse/CA-20